stable-requirements.txt
BeautifulSoup==3.2.1
Fabric==1.8.3
Flask==0.9
Flask-MySQL==1.2
IPy==0.75
Jinja2==2.6
MarkupSafe==0.19
MySQL-python==1.2.3
PyYAML==3.10
Pygments==1.5
SQLAlchemy==0.8.0
Twisted==13.2.0
Werkzeug==0.8.3
argparse==1.2.1
boto==2.3.0
cssselect==0.9.1
ecdsa==0.11
flexmock==0.9.7
google-api-python-client==1.1
gunicorn==0.17.2
httplib2==0.8
ipython==0.13.1
itsdangerous==0.23
logilab-astng==0.24.3
logilab-common==0.59.1
lxml==3.3.3
matplotlib==1.2.1
nose==1.1.2
openpyxl==1.8.0
oursql==0.9.3.1
pandas==0.11.0
paramiko==1.12.3
pexpect==2.4
premailer==1.13
protobuf==2.4.1
psycopg2==2.5.1
pyOpenSSL==0.13.1
pyasn1==0.1.6
pycrypto==2.6.1
pydot==1.0.2
pydot2==1.0.33
pyflakes==0.5.0
pylint==0.28.0
pyparsing==1.5.6
python-Levenshtein==0.10.2
python-graph-core==1.8.2
python-graph-dot==1.8.2
python-sshtail==0.0.2
pytz==2012c
requests==1.1.0
scipy==0.12.0
selenium==2.39.0
sendgrid==0.1.3
six==1.4.1
smtpapi==0.1.1
ssh==1.8.0
thrift==0.9.0
ua-parser==0.3.4
ujson==1.33
unittest2==0.5.1
urllib3==1.3
user-agents==0.1.1
vincent==0.1.4
virtualenv==1.9.1
wsgiref==0.1.2
xlrd==0.9.2
zope.interface==4.0.5
